extract allocation profiling parts
